+++ CHANGES.md
@@ -26,6 +26,7 @@ Changes in 11.0
 - Add Class Average & Class Rank (Course Period) in ReportCards.fnc.php
 - Move "Progress Reports" from Teacher Programs to Grades menu (admin) in Grades/Menu.php, ProgressReports.php, rosariosis.sql & rosariosis_mysql.sql
 - CSS FlatSIS enlarge left menu width by 6px for German in stylesheet.css
+- SQL add ORDER BY to GetChildrenMP() in GetMP.php
 
 Changes in 10.9.6
 -----------------